+++ metatensor-torch/include/metatensor/torch/atomistic/system.hpp
@@ -117,7 +117,7 @@ public:
         torch::autograd::AutogradContext* ctx,
         torch::Tensor positions,
         torch::Tensor cell,
-        TorchTensorBlock neighbors,
+        TensorBlock neighbors,
         bool check_consistency
     );
 
@@ -140,7 +140,7 @@ public:
 /// checks in case the data in neighbors does not follow what's expected.
 METATENSOR_TORCH_EXPORT void register_autograd_neighbors(
     System system,
-    TorchTensorBlock neighbors,
+    TensorBlock neighbors,
     bool check_consistency
 );
 
@@ -242,11 +242,11 @@ public:
     /// The neighbors values must contain the distance vector from the first to
     /// the second atom, i.e. `positions[second_atom] - positions[first_atom] +
     /// cell_shift_a * cell_a + cell_shift_b * cell_b + cell_shift_c * cell_c`.
-    void add_neighbor_list(NeighborListOptions options, TorchTensorBlock neighbors);
+    void add_neighbor_list(NeighborListOptions options, TensorBlock neighbors);
 
     /// Retrieve a previously stored neighbor list with the given options, or
     /// throw an error if no such neighbor list exists.
-    TorchTensorBlock get_neighbor_list(NeighborListOptions options) const;
+    TensorBlock get_neighbor_list(NeighborListOptions options) const;
 
     /// Get the options for all neighbor lists registered with this `System`
     std::vector<NeighborListOptions> known_neighbor_lists() const;
@@ -260,10 +260,10 @@ public:
     /// @param tensor the data to store
     /// @param override if true, allow this function to override existing data
     ///                 with the same name
-    void add_data(std::string name, TorchTensorMap tensor, bool override=false);
+    void add_data(std::string name, TensorMap tensor, bool override=false);
 
     /// Retrieve custom data stored in this System, or throw an error.
-    TorchTensorMap get_data(std::string name) const;
+    TensorMap get_data(std::string name) const;
 
     /// Get the list of data registered with this `System`
     std::vector<std::string> known_data() const;
@@ -292,8 +292,8 @@ private:
     torch::Tensor cell_;
     torch::Tensor pbc_;
 
-    std::map<NeighborListOptions, TorchTensorBlock, nl_options_compare> neighbors_;
-    std::unordered_map<std::string, TorchTensorMap> data_;
+    std::map<NeighborListOptions, TensorBlock, nl_options_compare> neighbors_;
+    std::unordered_map<std::string, TensorMap> data_;
 };
 
 }